Pass 3 Task 4: extract GhostPlaybackLogic from ParsekFlight
Move 54 static ghost playback methods + 3 constants from ParsekFlight
into new GhostPlaybackLogic static class (1498 lines). Eliminates
ParsekKSC->ParsekFlight method coupling entirely. ParsekFlight drops
from ~8500 to ~7000 lines. Fix bare Destroy() calls to
UnityEngine.Object.Destroy() for non-MonoBehaviour context. Update
call sites in ParsekFlight, ParsekKSC, ParsekScenario, and 6 test
files. Zero logic changes.
